% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_combinations.R
\name{all_correlations}
\alias{all_correlations}
\title{Compute correlations for all possible mask combinations given the input vectors.}
\usage{
all_correlations(vs, mask_ranges)
}
\arguments{
\item{vs}{Input vectors}

\item{mask_ranges}{List of start, end tuples for each of the mask groups}
}
\value{
A dictionary of mask combinations and the respective correlations.
}
\description{
Compute correlations for all possible mask combinations given the input vectors.
}

