% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_combinations.R
\name{all_pairwise_correlations}
\alias{all_pairwise_correlations}
\title{Given two vectors, compute correlations for all possible masks}
\usage{
all_pairwise_correlations(v1, v2, m1, m2, mask_ranges)
}
\arguments{
\item{v1}{Input vector 1}

\item{v2}{Input vector 2}

\item{m1}{Pre-computed mask for v1}

\item{m2}{Pre-computed mask for v2}

\item{mask_ranges}{List of start, end tuples for each of the mask groups}
}
\value{
A vector of correlations for all possible masks. If a given mask is in
 conflict with the input masks, the entry in the vector will contain \code{NA}
 in that position.
}
\description{
Given two vectors, compute correlations for all possible masks
}

