% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_combinations.R
\name{best_masked_correlation}
\alias{best_masked_correlation}
\title{Compute correlations for the largest possible mask given the input vectors,}
\usage{
best_masked_correlation(vs, mask_ranges)
}
\arguments{
\item{vs}{Input vectors}

\item{mask_ranges}{List of start, end tuples for each of the mask groups}
}
\value{
A list of the correlation and the encountered mask.
}
\description{
Compute correlations for the largest possible mask given the input vectors,
}

