% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contains_singleton}
\alias{contains_singleton}
\title{Test if any bit in the input masks is set exactly once}
\usage{
contains_singleton(vs, N_bits)
}
\description{
Test if any bit in the input masks is set exactly once
}

