% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_masks.R
\name{detect_mask}
\alias{detect_mask}
\title{Detect mask groups in a vector}
\usage{
detect_mask(v, mask_ranges)
}
\arguments{
\item{v}{Input vector}

\item{mask_ranges}{List of start, end tuples for each of the mask groups}
}
\value{
A number in which the i-th bit corresponds to the absence of NAs
 in the i-th mask group,
}
\description{
Detect mask groups in a vector
}
\examples{
detect_mask(c(NA,NA,3:10), list(c(1,2), c(3,6), c(7,10)))
}

