% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_operations.R
\name{increment_mask}
\alias{increment_mask}
\title{Cycle through all possible mask combinations by incrementing the mask}
\usage{
increment_mask(mask, N)
}
\arguments{
\item{mask}{Input mask}

\item{N}{threshold for setting one part of the mask to zero and moving to next}
}
\value{
The incremented mask, or \code{NA} if all combinations have been reached
}
\description{
Cycle through all possible mask combinations by incrementing the mask
}

