% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_to_masks}
\alias{int_to_masks}
\title{Split from integer value into masks}
\usage{
int_to_masks(int_masks, bits_per_mask, N_masks)
}
\arguments{
\item{int_masks}{encoded masks}

\item{bits_per_mask}{Number of bits in each mask}

\item{N_masks}{Number of masks}
}
\description{
Split from integer value into masks
}

