% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{masked_corr}
\alias{masked_corr}
\title{Calculate correlation only for ranges where the mask bit has been set}
\usage{
masked_corr(v1, v2, mask, mask_ranges)
}
\arguments{
\item{v1}{Input vector 1}

\item{v2}{Input vector 2}

\item{mask}{Mask, each bit that is set will be included in the correlation}

\item{mask_ranges}{List of start, end tuples for each of the mask groups}
}
\value{
The Pearson correlation between the designated parts of the input vectors
}
\description{
Calculate correlation only for ranges where the mask bit has been set
}

