% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{masks_to_int}
\alias{masks_to_int}
\title{Convert the mask to an integer for easier handling with dplyr
NOTE! We have at most 6 elements in the mask, and 4 bits = 24 bits.
so we're safe.}
\usage{
masks_to_int(masks, bits_per_mask)
}
\arguments{
\item{masks}{Integer vector of masks}

\item{bits_per_mask}{Number of bits in each mask}
}
\description{
Convert the mask to an integer for easier handling with dplyr
NOTE! We have at most 6 elements in the mask, and 4 bits = 24 bits.
so we're safe.
}

