% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{masks_to_target}
\alias{masks_to_target}
\title{Generate vector of masks containing all intersections of bits}
\usage{
masks_to_target(masks)
}
\arguments{
\item{masks}{input set of masks}
}
\value{
same as upper.tri of outer(masks, masks), bitwAnd)
}
\description{
Generate vector of masks containing all intersections of bits
}

