% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_combinations.R
\name{which.most_set_bits}
\alias{which.most_set_bits}
\title{Find mask with most keys set}
\usage{
which.most_set_bits(masks)
}
\arguments{
\item{masks}{List of masks}
}
\value{
The items with the most set bits. Will raise an error if more
 than one items have the same number of set bits
}
\description{
Find mask with most keys set
}

