% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{increment_mask}
\alias{increment_mask}
\title{Cycle through all possible mask combinations by incrementing the mask}
\usage{
increment_mask(mask, N_bits, skip_singletons, skip_zeros)
}
\arguments{
\item{mask}{Input mask}

\item{skip_singletons}{If set, do not return singletons}

\item{skip_zeros}{If set, do not return masks where any submask is 0}

\item{N}{threshold for setting one part of the mask to zero and moving to next}
}
\value{
The incremented mask, or \code{NA} if all combinations have been reached
}
\description{
Cycle through all possible mask combinations by incrementing the mask
}

