% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_combinations.R
\name{mean_correlation_for_mask_set}
\alias{mean_correlation_for_mask_set}
\title{Given a precomputed matrix of masked correlations, gather correlations
for the specified mask set.}
\usage{
mean_correlation_for_mask_set(masks, correlations, Nbits, masks_filter = NULL)
}
\arguments{
\item{masks}{vector of masks (length N) from which to take correlations}

\item{correlations}{NxN matrix of correlations, where each matrix element is a vector
with entries for each masked correlation}

\item{Nbits}{optional: number of bits, must be supplied if masks_filter is given}

\item{masks_filter}{optional: set of masks (of the comparison, not the input mask) for which
to return correlations, supply as integer-encoded mask (not as vector)}
}
\value{
A list of two elements: the average correlation, and a sorted vector
 of the encountered combination of masks. It is possible that the given masks and
 correlations are partially not compatible, Then, the average will be computed
 based on those mask combinations that contain a correlation. (E.g. in the mask triple
 "A, AB, B" the average will be computed from two correlations (A/A, B/B), as it is
 not possible to calculate a correlation between A and B.)
}
\description{
Given a precomputed matrix of masked correlations, gather correlations
for the specified mask set.
}

